// frontend/admin/src/lib/sdk/models/app__api__admin__endpoints__chat_sessions__ChatMessage.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 会话消息
 */
export type app__api__admin__endpoints__chat_sessions__ChatMessage = {
    /**
     * 消息角色: user, assistant, system
     */
    role: string;
    /**
     * 消息内容
     */
    content: string;
};